Check in a Makefile that includes automatic provisioning of the venv.
There was already a Makefile in local be folks locally - extend this with
tricks to create the virtual env and install any requisite dependencies
within it. The added targets should also notice future dependency changes.

While here also add setuptools as an explicit dependency in order to allow
operation on Pyhton 3.12+ where it is no longer an included battery.