Apply a minimal set of changes/tweaks to restore a responsive narrow view.
Playing with disabling the custom CSS we apply lead to discovering that
the theme we use already had support for a collapsed sidebar view based
on a media query that came into effect at 768px.

Arrange applying essentially the same styles at our existing break point
of 1100px, remove a couple of conflicting things in our custom CSS and set
a handful of basic styles (in line with choices already made in the page)
such that the responsive view comes into a working order.